# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: متجر الأفق المتجدد API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories/{slug}:
    get:
      operationId: getCategory
      tags: [catalog]
      summary: Get a category by slug
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Category
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CategoryWithSubs" }
  /subcategories/{slug}/products:
    get:
      operationId: listSubcategoryProducts
      tags: [catalog]
      summary: List products in a subcategory
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
        - in: query
          name: sort
          schema:
            type: string
            enum: [price_asc, price_desc, newest]
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                type: object
                required: [subcategory, products]
                properties:
                  subcategory: { $ref: "#/components/schemas/Subcategory" }
                  category: { $ref: "#/components/schemas/Category" }
                  products:
                    type: array
                    items: { $ref: "#/components/schemas/Product" }
  /products/{id}:
    get:
      operationId: getProduct
      tags: [catalog]
      summary: Get a product by id
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
...
  /orders:
    get:
      operationId: listMyOrders
      tags: [orders]
      summary: List the current user's orders
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Order" }
    post:
      operationId: placeOrder
      tags: [orders]
      summary: Place an order from the current cart
      parameters:
        - $ref: "#/components/parameters/SessionToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/PlaceOrderBody" }
      responses:
        "200":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get a specific order
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /orders/{id}/cancel:
    post:
      operationId: cancelOrder
      tags: [orders]
      summary: Cancel my order
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
  /orders/{id}/resolve-stockout:
    post:
      operationId: resolveStockout
      tags: [orders]
      summary: Customer decision after a stockout (continue without item or cancel order)
      parameters:
        - $ref: "#/components/parameters/SessionToken"
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ResolveStockoutBody" }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
...
components:
  parameters:
    SessionToken:
      in: header
      name: x-session-token
      required: false
      schema: { type: string }
  schemas:
...
    Category:
      type: object
      required: [id, slug, name]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        nameEn: { type: string, nullable: true }
        icon: { type: string, nullable: true }
        imageUrl: { type: string, nullable: true }
    Subcategory:
      type: object
      required: [id, slug, name, categoryId]
      properties:
        id: { type: integer }
        slug: { type: string }
        name: { type: string }
        nameEn: { type: string, nullable: true }
        categoryId: { type: integer }
    SubcategoryWithCount:
      allOf:
        - $ref: "#/components/schemas/Subcategory"
        - type: object
          properties:
            productCount: { type: integer }
    CategoryWithSubs:
      allOf:
        - $ref: "#/components/schemas/Category"
        - type: object
          required: [subcategories]
          properties:
            subcategories:
              type: array
              items: { $ref: "#/components/schemas/SubcategoryWithCount" }
    ProductColor:
      type: object
      required: [id, productId, name, hexCode, stock, enabled, sortOrder]
      properties:
        id: { type: integer }
        productId: { type: integer }
        name: { type: string }
        hexCode: { type: string }
        stock: { type: integer }
        enabled: { type: boolean }
        sortOrder: { type: integer }
    Product:
      type: object
      required: [id, name, price, description, imageUrl, stock, subcategoryId, subcategorySlug, subcategoryName, categorySlug, categoryName]
      properties:
        id: { type: integer }
        name: { type: string }
        price: { type: number }
        description: { type: string }
        imageUrl: { type: string }
        image2: { type: string, nullable: true }
        image3: { type: string, nullable: true }
        colorsEnabled: { type: boolean }
        stock: { type: integer }
        subcategoryId: { type: integer }
        subcategorySlug: { type: string }
        subcategoryName: { type: string }
        categorySlug: { type: string }
        categoryName: { type: string }
        createdAt: { type: string }
        averageRating: { type: number, nullable: true }
        ratingCount: { type: integer }
        colors:
          type: array
          items: { $ref: "#/components/schemas/ProductColor" }
...
    OrderStatus:
      type: string
      description: "حالة الطلب"
      enum: [pending, preparing, prepared, delivered_to_courier, cancelled, stockout, payment_received]
    PlaceOrderBody:
      type: object
      required: [region, address, phone, backupPhone]
      properties:
        region: { type: string }
        address: { type: string }
        phone: { type: string }
        backupPhone: { type: string }
        notes: { type: string }
        deliveryFee: { type: number }
...
    OrderItem:
      type: object
      required: [id, productId, productName, productImage, price, quantity, outOfStock]
      properties:
        id: { type: integer }
        productId: { type: integer }
        productName: { type: string }
        productImage: { type: string }
        price: { type: number }
        quantity: { type: integer }
        outOfStock: { type: boolean }
        selectedColor: { type: string, nullable: true }
    Order:
      type: object
      required: [id, userId, status, total, region, address, phone, backupPhone, items, createdAt]
      properties:
        id: { type: integer }
        userId: { type: integer }
        userName: { type: string }
        status: { $ref: "#/components/schemas/OrderStatus" }
        total: { type: number }
        deliveryFee: { type: number }
        region: { type: string }
        address: { type: string }
        phone: { type: string }
        backupPhone: { type: string }
        notes: { type: string, nullable: true }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        createdAt: { type: string }
...
    ResolveStockoutBody:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [continue_without, cancel]
...