# lib/api-zod/src/generated/types/resolveStockoutBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * متجر الأفق المتجدد API
 * OpenAPI spec version: 0.1.0
 */
import type { ResolveStockoutBodyAction } from "./resolveStockoutBodyAction";

export interface ResolveStockoutBody {
  action: ResolveStockoutBodyAction;
}